Add files via upload
The Credit Card Fraud Detection Problem includes modeling past credit card transactions with the knowledge 
of the ones that turned out to be fraud. This model is then used to identify whether a new transaction 
is fraudulent or not. Our aim here is to detect 100% of the fraudulent transactions while minimizing the 
incorrect fraud classifications.